finish moving circuit runners
