+++ src/lib/proof_system.ts
@@ -465,7 +465,7 @@ function analyzeMethod<T>(
   methodIntf: MethodInterface,
   method: (...args: any) => T
 ) {
-  return Circuit.constraintSystem(() => {
+  return Provable.constraintSystem(() => {
     let args = synthesizeMethodArguments(methodIntf, true);
     let publicInput = emptyWitness(publicInputType);
     return method(publicInput, ...args);
